requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-filter
django_extensions
requests
Pillow
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
mssql-django==1.4
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
whitenoise==6.7.0
Faker==25.2.0

# Azure dependencies (optional - will fallback gracefully if not available)
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
applicationinsights==0.11.10
